update analyze loops to include egge distance and length

+++ graph_analysis/link_dataset.py
@@ -328,7 +328,7 @@ class LinkVesselGraph(InMemoryDataset):
 
             # includes all edges (train+test+val) in both drections
             data.edge_index_undirected = torch.tensor(edge_array, dtype=torch.long)
-            data.edge_attr_undirected = torch.tensor(edge_attr_array, dtype=torch.long)
+            data.edge_attr_undirected = torch.tensor(edge_attr_array, dtype=torch.float)
             del edge_array
             del edge_attr_array
